Moved the files for the Bluetooth and Wi-Fi thunder plugins(bluetoothManager.js and wifiManager.js) to folder named rdkservices.
